requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==41.0.7
DateTime==5.5
Faker==37.3.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
-e .
lxml==4.9.4
MarkupSafe==3.0.2
multidict==6.4.4
openleadr==0.5.35
packaging==25.0
pluggy==1.6.0
propcache==0.3.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyOpenSSL==23.2.0
pytest==8.3.5
python-dotenv==1.1.0
pytz==2025.2
setuptools==80.9.0
signxml==3.2.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.2
xmltodict==0.13.0
yarl==1.20.0
zope.interface==7.2
